refactor(rust): splitter external_imap/mod.rs (694→196) et mailbox/mod.rs (527→23)
external_imap/mod.rs : 694 → 196 LOC
- Extraction du client IMAP + commandes + parser vers sous-modules
- mod.rs devient re-exports uniquement

email_api_dir/mailbox/mod.rs : 527 → 23 LOC
- mime_utils.rs (448 LOC) : helpers MIME/parsing
- folder_utils.rs (68 LOC) : helpers folder
- mod.rs : re-exports uniquement

Aucun changement de comportement.
